requirements.txt
Flask[async]
PyMySQL
requests
python-dotenv
sqlalchemy[asyncio]
aiomysql
hypercorn
backoff